requirements.txt
beautifulsoup4==4.12.2
bencoder.pyx==3.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
feedparser==6.0.10
idna==3.4
qbittorrent-api==2023.5.48
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.2
